refactor: extract boundary-aware redirect logic into shared boundaryUtils
- buildContainerBoundaryMap and findPublicApiCandidate moved to boundaryUtils.ts
- resolveRedirectTarget added: same-boundary → repo, cross-boundary → public API
- fixDbPerService and fixCrud both use shared utils
- fixCrud Type 1 is now boundary-aware: cross-boundary with no existing repo
  warns and skips instead of creating a repo in the wrong context
