+++ src/rules/fixDbPerService.ts
@@ -1,7 +1,11 @@
 import consola from "consola";
 
-import type { ArchitectureModel, Boundary, Container } from "../model";
+import type { ArchitectureModel, Container } from "../model";
 import { CONTAINER_DB_TYPE } from "../model";
+import {
+  buildContainerBoundaryMap,
+  resolveRedirectTarget,
+} from "./boundaryUtils";
 import type { DbPerServiceOptions } from "./dbPerService";
 import type { FixResult, SourceSyntax } from "./fix";
 import type { Violation } from "./types";
@@ -31,55 +35,6 @@ const resolveOwner = (
   return tagged[0];
 };
 
-const buildContainerBoundaryMap = (
-  model: ArchitectureModel,
-): Map<string, Boundary> => {
-  const map = new Map<string, Boundary>();
-  for (const boundary of model.boundaries) {
-    for (const container of boundary.containers) {
-      map.set(container.name, boundary);
-    }
-  }
-  return map;
-};
-
-/**
- * Finds the best "public API" container in a boundary to serve as redirect
- * target for cross-boundary accessors. Prefers containers with the most
- * incoming relations from outside the boundary (highest in-degree).
- */
-const findPublicApiCandidate = (
-  targetBoundary: Boundary,
-  dbType: string,
-  ownerTags: string[],
-  model: ArchitectureModel,
-  containerBoundaryMap: Map<string, Boundary>,
-): Container | undefined => {
-  const candidates = targetBoundary.containers.filter(
-    (c) => c.type !== dbType && !ownerTags.some((t) => c.tags?.includes(t)),
-  );
-
-  if (candidates.length === 0) return undefined;
-  if (candidates.length === 1) return candidates[0];
-
-  // Rank candidates by in-degree from containers outside this boundary
-  const candidateNames = new Set(candidates.map((c) => c.name));
-  const inDegree = new Map<string, number>(candidates.map((c) => [c.name, 0]));
-
-  for (const container of model.allContainers) {
-    if (containerBoundaryMap.get(container.name) === targetBoundary) continue;
-    for (const rel of container.relations) {
-      if (candidateNames.has(rel.to.name)) {
-        inDegree.set(rel.to.name, (inDegree.get(rel.to.name) ?? 0) + 1);
-      }
-    }
-  }
-
-  return candidates.toSorted(
-    (a, b) => (inDegree.get(b.name) ?? 0) - (inDegree.get(a.name) ?? 0),
-  )[0];
-};
-
 export const fixDbPerService = (
   model: ArchitectureModel,
   violations: Violation[],
@@ -103,7 +58,6 @@ export const fixDbPerService = (
     if (accessors.length <= 1) continue;
 
     const owner = resolveOwner(db.name, accessors, ownerTags);
-    const dbBoundary = containerBoundaryMap.get(db.name);
 
     const edits = accessors
       .filter((c) => c !== owner)
@@ -116,38 +70,17 @@ export const fixDbPerService = (
           return [];
         }
 
-        const accessorBoundary = containerBoundaryMap.get(accessor.name);
-        const isCrossBoundary =
-          accessorBoundary !== undefined &&
-          dbBoundary !== undefined &&
-          accessorBoundary !== dbBoundary;
-
-        let redirectTarget: Container;
-
-        if (isCrossBoundary) {
-          const publicApi = findPublicApiCandidate(
-            dbBoundary,
-            dbType,
-            ownerTags,
-            model,
-            containerBoundaryMap,
-          );
-          if (!publicApi) {
-            consola.warn(
-              `fix dbPerService: boundary "${dbBoundary.name}" has no public API — cannot auto-redirect "${accessor.name}" away from "${db.name}", fix manually`,
-            );
-            return [];
-          }
-          if (publicApi === owner) {
-            consola.warn(
-              `fix dbPerService: the only public API candidate in "${dbBoundary.name}" is the repo owner — cross-boundary access from "${accessor.name}" requires manual review`,
-            );
-            return [];
-          }
-          redirectTarget = publicApi;
-        } else {
-          redirectTarget = owner;
-        }
+        const redirectTarget = resolveRedirectTarget(
+          accessor,
+          db,
+          owner,
+          dbType,
+          ownerTags,
+          model,
+          containerBoundaryMap,
+          "dbPerService",
+        );
+        if (!redirectTarget) return [];
 
         const tags =
           rel.tags && rel.tags.length > 0 ? rel.tags.join("+") : undefined;
